# backend/src/API.yml
openapi: 3.0.3
info:
  description: This is the interactive documentation for the Innovatiehuis Intranet RESTful API
  version: "0.0.0"
  title: Innovatieplatform REST API
servers:
  - url: http://127.0.0.1:5000/api



  # Paths supported by the server application
paths:
...
  /projects/{project_id}:
    get:
      operationId: src.endpoints.projects.read_one
      tags:
        - Projects
      summary: Read one project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read project operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      operationId: src.endpoints.projects.update
      tags:
        - Projects
      summary: Update a value of project
      parameters:
        - $ref: '#/components/parameters/project_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                description:
                  type: string

      responses:
        200:
          description: Project updated
    delete:
      operationId: src.endpoints.projects.delete
      tags:
        - Projects
      summary: Deletes the specified project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Project deleted
  /projects/{project_id}/archive:
    put:
      operationId: src.endpoints.projects.update_archive
      tags:
        - Projects
      summary: Archive or de-archive a project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Project archive status updated
  /projects/{project_id}/parents:
    get:
      operationId: src.endpoints.projects.read_parents
      tags:
        - Projects
      summary: Read all projects marked as parent of a specific project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read all parent projects for this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{project_id}/parents/{parent_id}:
    post:
      operationId: src.endpoints.projects.add_parent
      tags:
        - Projects
      summary: Read all projects marked as parent of a specific project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/parent_id'
      responses:
        200:
          description: Successfully assigned parent project to given project
    delete:
      operationId: src.endpoints.projects.remove_parent
      tags:
        - Projects
      summary: Read all projects marked as parent of a specific project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/parent_id'
      responses:
        200:
          description: Successfully broke parent-child relation
  /projects/{project_id}/children:
    get:
      operationId: src.endpoints.projects.read_children
      tags:
        - Projects
      summary: Read all projects marked as child of a specific project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read child projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{project_id}/children/{child_id}:
    post:
      operationId: src.endpoints.projects.add_child
      tags:
        - Projects
      summary: Mark a project as child to the given project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/child_id'
      responses:
        200:
          description: Successfully marked project as child
    delete:
      operationId: src.endpoints.projects.remove_child
      tags:
        - Projects
      summary: Remove a project as child from a given project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/child_id'
      responses:
        200:
          description: Successfully broke parent-child relation
  /projects/{project_id}/children/{child_id}/shared_files:
    put:
      operationId: src.endpoints.projects.update_shared_files
      tags:
        - Projects
        - Files
      summary: Update what files are shared from parent project to child project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/child_id'
      responses:
        200:
          description: Successfully updated shared files
  /projects/{project_id}/users:
    get:
      operationId: src.endpoints.projects.read_users
      tags:
        - Projects
        - Users
      summary: Get IDs of all users assigned to this project, as well as the time they last accessed it
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read users assigned to this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      operationId: src.endpoints.projects.update_users
      tags:
        - Projects
        - Users
      summary: Change the users in this project by supplying a list of user IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IDs'
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully updated list of users assigned to this project
  /projects/{project_id}/users/{user_id}:
    post:
      operationId: src.endpoints.projects.add_user
      tags:
        - Projects
        - Users
      summary: Add a user to a project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully added user to project
    delete:
      operationId: src.endpoints.projects.remove_user
      tags:
      - Projects
      - Users
      summary: Remove a user from a project
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully removed user from project
  /projects/{project_id}/users/last_seen:
    patch:
      operationId: src.endpoints.projects.update_last_seen
      tags:
        - Projects
        - Users
      summary: Lets the data model know when a user has last accessed the contents of a Project.
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully updated the time when the given User last saw this Project.
  /projects/{project_id}/announcements:
    get:
      operationId: src.endpoints.projects.read_announcements
      tags:
      - Projects
      - Announcements
      summary: Get all announcements from project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully retrieved announcements from project
        404:
          description: Specified project not found
    post:
      operationId: src.endpoints.projects.add_announcement
      tags:
      - Projects
      - Announcements
      summary: Post announcements in project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully posted announcement to project

# ========================================FILES/FOLDERS============================================
...
  /announcements/{announcement_id}:
    get:
      operationId: src.endpoints.announcements.read_one
      tags:
        - Announcements
      summary: Reads a specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully retrieved single announcement
    put:
      operationId: src.endpoints.announcements.update
      tags:
        - Announcements
      summary: Update title and/or content of a specific announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully posted announcement to project
    delete:
      operationId: src.endpoints.announcements.delete
      tags:
        - Announcements
      summary: Deletes specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully removed announcement
...
  /users/{user_id}:
    get:
      operationId: src.endpoints.users.read_one
      tags:
        - Users
      summary: Get specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully read user
    put:
      operationId: src.endpoints.users.update
      tags:
        - Users
      summary: Edit user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User' # BUT NOT PASSWORD_HASH
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully edited user
        403:
          description: Request denied due to lack of permissions
    delete:
      operationId: src.endpoints.users.delete
      tags:
        - Users
      summary: Delete specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully deleted user
...
components:
  parameters:
    project_id:
      in: path
      name: project_id
      required: true
      schema:
        type: string
      description: The ID number of a project
    parent_id:
      in: path
      name: parent_id
      required: true
      schema:
        type: string
      description: The ID number of a parent project
    child_id:
      in: path
      name: child_id
      required: true
      schema:
        type: string
      description: The ID number of a child project
    user_id:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      description: The ID number of a user
    announcement_id:
      in: path
      name: announcement_id
      required: true
      schema:
        type: string
      description: The ID number of an announcement
...
  schemas:
    Project:
      type: object
      properties:
        projectid:
          type: integer
        project_name:
          type: string
        is_archived:
          type: boolean
        create:
          type: string
        last_updated:
          type: string
    User:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone_number:
          type: string
        roleid:
          type: integer
        screening_status:
          type: integer
        password_hash:
          type: string
    Announcement:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
...
    IDs:
      type: object
      properties:
        IDs:
          type: array